requirements.txt
flask==3.0.3
ultralytics==8.3.195
torch==2.0.1
torchvision==0.15.2
opencv-python>=4.1.1
pillow==10.2.0
numpy<2.0